Added to the D-n-D Function
I created a new file, called shipPanels, which will hold the class panels for all of the five different ships. I added code to this file that *should* snap the label into place with the grid. Methods for adding this data to the model and getting the grid size is still needed. I haven't been able to test it.