Slight modifications to view and Began server implementation
Modified the view so red or gray dots appear when shooting. Also began implementing the server side, however still need to add the client and integrate interaction with the controller (via battleshipServer class). I also added an extra folder, that's just for me to make it easier. ill delete it once I'm done.